Fixes to previous commit
Replaced worldPointMap to add back water animation tiles and remove
curruption artifacts that encompassed ~2/3 of the file.

Minor code fixes in order to make the commit compilable. Note that one
of these was to check wether the audioClip is null before trying to stop
the thread.

Bugs that need to be addressed:
- There is a lagged delay in playing the door transiton clip and wall
bump clip.
- Walking cycle should run every time the wall bump clip is played.
- Battle music need only be played when a battle even is triggered (I
removed the tiles that triggered this in the worldPointMap and will
instead be implemented with the battle system.
- Battle music has a random pokemon sound in it needs to be removed.
- Intro song never plays because the audioClip never starts (hence the
check for nullPointer on stopping an audio thread)